more proof of concept work to determine which library to use for parse compare
