+++ Python Console Project/ChunkfilePandas.py
@@ -4,7 +4,7 @@ import os
 import pandas as pd
 import datetime
 
-def split_csv_file(input_file, output_folder, max_rows=2000000):
+def split_csv_file(input_file, output_folder, max_rows=1000):
     # Create the output folder if it doesn't exist
     if not os.path.exists(output_folder):
         os.makedirs(output_folder)
@@ -22,14 +22,14 @@ def split_csv_file(input_file, output_folder, max_rows=2000000):
 
         if rows_count == max_rows:
             # Create the output file path
-            output_file = os.path.join(output_folder, f"chunk_{chunk_count}.csv")
+            #output_file = os.path.join(output_folder, f"chunk_{chunk_count}.csv")
 
             # Create a DataFrame from the accumulated rows
             chunk_df = pd.DataFrame(chunk_rows, columns=reader.columns)
 
             # Write the chunk data to the output file
-            chunk_df.to_csv(output_file, index=False)
-            print(f"Chunk {chunk_count}: {output_file} created")
+            #chunk_df.to_csv(output_file, index=False)
+            #print(f"Chunk {chunk_count}: {output_file} created")
 
             chunk_count += 1
             rows_count = 0
@@ -37,10 +37,10 @@ def split_csv_file(input_file, output_folder, max_rows=2000000):
 
     # Write the remaining chunk to a file
     if rows_count > 0:
-        output_file = os.path.join(output_folder, f"chunk_{chunk_count}.csv")
+        #output_file = os.path.join(output_folder, f"chunk_{chunk_count}.csv")
         chunk_df = pd.DataFrame(chunk_rows, columns=reader.columns)
-        chunk_df.to_csv(output_file, index=False)
-        print(f"Chunk {chunk_count}: {output_file} created")
+        #chunk_df.to_csv(output_file, index=False)
+        #print(f"Chunk {chunk_count}: {output_file} created")
 
     print("File splitting completed!")
