+++ proof-of-concept/digital-preservation-poc/pom.xml
@@ -16,7 +16,7 @@
 	        <dependency>
 	            <groupId>uk.gov.nationalarchives</groupId>
 	            <artifactId>csv-validator-java-api</artifactId>
-	            <version>1.1.5</version>
+	            <version>1.2-RC4</version>
 	<!--             <exclusions> -->
 	<!--                 <exclusion> -->
 	<!--                     <groupId>com.gilt</groupId> -->
@@ -24,12 +24,12 @@
 	<!--                 </exclusion> -->
 	<!--             </exclusions> -->
 	</dependency>
-	<dependency>
 	<!-- required for nationalarchives jar above-->
-	          <groupId>com.gilt</groupId>
-	          <artifactId>gfc-semver_2.11</artifactId>
-	          <version>0.0.5</version>
-	      </dependency>
+<!--	<dependency>-->
+<!--	          <groupId>com.gilt</groupId>-->
+<!--	          <artifactId>gfc-semver_2.11</artifactId>-->
+<!--	          <version>0.0.5</version>-->
+<!--	      </dependency>-->
 	</dependencies>
 
 </project>
