+++ proof-of-concept/parse-compare/.gitignore
@@ -1,0 +1,1 @@
+/target/
